+++ grain/_src/python/dataset/dataset.py
@@ -741,17 +741,25 @@ class MapDataset(_Dataset, Generic[T], metaclass=MapDatasetMeta):
         parent=self, transform=transform, seed=seed
     )
 
-  def repeat(self, num_epochs: int | None = None) -> MapDataset[T]:
+  def repeat(
+      self, num_epochs: int | None = None, *, reseed_each_epoch: bool = True
+  ) -> MapDataset[T]:
     """Returns a dataset repeating the elements of this dataset multiple times.
 
     Specifying ``None`` for ``num_epochs`` will repeat the dataset infinitely,
     and causes ``len(ds)`` to return ``sys.maxsize``.
 
-    Since ``MapDataset`` allows accessing elements past ``len(ds) - 1`` anyway
-    (and uses the index modulo ``len(ds)``), this transformation effectively
-    only changes the length of the dataset.
-
     Can not be called on an infinite dataset.
+
+    By default, random upstream transformations such as ``shuffle`` and
+    ``random_map`` will be seeded differently (but deterministically) for each
+    epoch. This means that each epoch will be shuffled differently. If you want
+    to repeat the first epoch exactly as is, set ``reseed_each_epoch`` to
+    ``False``.
+
+    Implementation note: since ``MapDataset`` allows accessing elements past
+    ``len(ds) - 1`` anyway (and uses the index modulo ``len(ds)``), this
+    transformation effectively only changes the length of the dataset.
 
     Example usage::
 
@@ -764,6 +772,10 @@ class MapDataset(_Dataset, Generic[T], metaclass=MapDatasetMeta):
     Args:
       num_epochs: Either a positive integer representing the number of times
         this dataset should be repeated or ``None`` to repeat infinitely.
+      reseed_each_epoch: If True (default), all random upstream transformations
+        use a seed folded with the epoch number, so that each epoch is shuffled
+        and processed differently. If False, the first epoch is repeated exactly
+        as is. In both cases, the random transformations are deterministic.
 
     Returns:
       A dataset repeating the elements of this dataset multiple times.
@@ -772,7 +784,11 @@ class MapDataset(_Dataset, Generic[T], metaclass=MapDatasetMeta):
     # pylint: disable=g-import-not-at-top
     from grain._src.python.dataset.transformations import repeat
     # pylint: enable=g-import-not-at-top
-    return repeat.RepeatMapDataset(parent=self, num_epochs=num_epochs)
+    return repeat.RepeatMapDataset(
+        parent=self,
+        num_epochs=num_epochs,
+        reseed_each_epoch=reseed_each_epoch,
+    )
 
   def to_iter_dataset(
       self,